Implemented heuristic and added test
- finished implementation of heuristic
- added unmovability table to rank goals
- implemented bbsolver for assignment of blocks to goals
- added tests

+++ src/manhattandist.cpp
@@ -16,7 +16,7 @@
 ManhattanDist :: ManhattanDist (Graph &g, Coord goal){
     graph = g;
     this->goal = goal;
-    
+
     computeDist();
 }
 
@@ -29,41 +29,41 @@ ManhattanDist :: ~ManhattanDist(){
 void ManhattanDist :: computeDist(){
     size_t len = State::height * State::width;
     dist = new int [len];
-  
+
     // Initialize all the squares to -1 (NOT_VISITED)
-    for (int i=0 ; i < len; i++) {
+    for (size_t i=0 ; i < len; i++) {
         dist [i] = NOT_VISITED;
     }
     std::queue <Coord> fringe;
     fringe.push (goal);
-    
+
     int d = 0;
     while (!fringe.empty ())
     {
         std::queue<Coord> newFringe;
-        
+
         while (!fringe.empty ()){
             Coord c = fringe.front ();
             fringe.pop();
-            
+
             if(dist[c.y * State::width + c.x] >= 0)
                 continue;
-            
+
             dist [c.y * State::width + c.x] = d;
-            
+
             std::vector <Coord> neighbors = graph.getNeighbors (c.x, c.y);
-            
-            for (int i=0; i< neighbors.size (); i++) {
+
+            for (size_t i=0; i< neighbors.size (); i++) {
                 Coord n = neighbors [i];
                 if (dist[n.y * State::width + n.x] == NOT_VISITED)
                     newFringe.push (n);
             }
         }
-        
-        
+
+
         fringe = newFringe;
         d++;
     }
-    
+
     // implement wavefront
 }
